move GenesisPfxInfo from chain to its own module in consensus
