+++ src/chain/mod.rs
@@ -17,45 +17,3 @@ pub use self::{
     chain::{delivery_group_size, Chain, ParsecResetData, PollAccumulated, SectionKeyShare},
     config::NetworkParams,
 };
-
-use crate::{
-    section::{AgeCounter, EldersInfo},
-    PublicId,
-};
-use std::{
-    collections::BTreeMap,
-    fmt::{self, Debug, Formatter},
-};
-
-#[derive(Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord, Hash, Clone)]
-pub struct GenesisPfxInfo {
-    pub elders_info: EldersInfo,
-    pub public_keys: bls::PublicKeySet,
-    pub state_serialized: Vec<u8>,
-    pub ages: BTreeMap<PublicId, AgeCounter>,
-    pub parsec_version: u64,
-}
-
-impl GenesisPfxInfo {
-    pub fn trimmed(&self) -> Self {
-        Self {
-            elders_info: self.elders_info.clone(),
-            public_keys: self.public_keys.clone(),
-            state_serialized: Vec::new(),
-            ages: self.ages.clone(),
-            parsec_version: self.parsec_version,
-        }
-    }
-}
-
-impl Debug for GenesisPfxInfo {
-    fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
-        write!(
-            formatter,
-            "GenesisPfxInfo({:?}, elders_version: {}, parsec_version {})",
-            self.elders_info.prefix(),
-            self.elders_info.version(),
-            self.parsec_version,
-        )
-    }
-}
